+++ src/test/TestStepList.cpp
@@ -11,12 +11,12 @@ TEST(TestStepList, TestList)
 {
     WW::StepList list;
 
-    ASSERT_EQ(0, list.size());
+    ASSERT_EQ(static_cast<size_t>(0), list.size());
 
     WW::TestStep step;
     step.short_desc("TestCase");
     list.push_back(&step);
-    ASSERT_EQ(1, list.size());
+    ASSERT_EQ(static_cast<size_t>(1), list.size());
     WW::StepList::const_iterator it = list.begin();
     ASSERT_EQ("TestCase", it->short_desc());
     ASSERT_NE(list.end(), it);
